Further work on tests
