+++ lib/oauth/client-credentials.js
@@ -107,7 +107,8 @@ OAuthClientCredentialsRequestAuthenticator.prototype._formatRequestBody = functi
 
   return {
     client_id: data.apiKey.id,
-    client_secret: data.apiKey.secret
+    client_secret: data.apiKey.secret,
+    grant_type: 'client_credentials'
   };
 };
 
@@ -161,8 +162,6 @@ OAuthClientCredentialsRequestAuthenticator.prototype.authenticate = function aut
 
   var href = application.href + '/oauth/token';
   var apiData = this._formatRequestBody(data);
-
-  apiData.grant_type = 'client_credentials';
 
   application.dataStore.createResource(href, {form: apiData}, function(err, data) {
     if (err) {
