+++ lib/stormpath.js
@@ -4,6 +4,7 @@ var passwordGrant = require('./oauth/password-grant');
 var refreshGrant = require('./oauth/refresh-grant');
 var idSiteGrant = require('./oauth/id-site-grant');
 var stormpathToken = require('./oauth/stormpath-token');
+var clientCredentialsGrant = require('./oauth/client-credentials');
 
 module.exports = {
   ApiKey: require('./authc/ApiKey'),
@@ -18,6 +19,8 @@ module.exports = {
   OAuthIdSiteTokenGrantAuthenticationResult: idSiteGrant.authenticationResult,
   OAuthStormpathTokenAuthenticator: stormpathToken.authenticator,
   OAuthStormpathTokenAuthenticationResult: stormpathToken.authenticationResult,
+  OAuthClientCredentialsAuthenticator: clientCredentialsGrant.authenticator,
+  OAuthClientCredentialsAuthenticationResult: clientCredentialsGrant.authenticationResult,
   SamlIdpUrlBuilder: require('./saml/SamlIdpUrlBuilder'),
   AssertionAuthenticationResult: require('./authc/AssertionAuthenticationResult'),
   StormpathAssertionAuthenticator: require('./authc/StormpathAssertionAuthenticator'),
